+++ CHANGES.rst
@@ -3,3 +3,5 @@ Changelog
 
 1.0.0 (unreleased)
 ------------------
+
+- #1  Hide unused fields from Analysis Request Add form